chore(telemetry/test): move testagent fixture to global conftest
