+++ tests/conftest.py
@@ -1,6 +1,9 @@
 import ast
+import base64
 import contextlib
+import importlib
 from itertools import product
+import json
 import os
 from os.path import split
 from os.path import splitext
@@ -8,21 +11,33 @@ import subprocess
 import sys
 from tempfile import NamedTemporaryFile
 import time
+from typing import Any
+from typing import Generator
+from typing import Tuple
 
 from _pytest.runner import call_and_report
 from _pytest.runner import pytest_runtest_protocol as default_pytest_runtest_protocol
+import attr
 import pytest
-from six import PY2
 
 import ddtrace
+from ddtrace.internal.compat import httplib
+from ddtrace.internal.compat import parse
 from ddtrace.internal.remoteconfig.client import RemoteConfigClient
 from ddtrace.internal.remoteconfig.worker import remoteconfig_poller
+from ddtrace.internal.telemetry import TelemetryWriter
+from ddtrace.internal.utils.formats import parse_tags_str
 from tests import utils
 from tests.utils import DummyTracer
 from tests.utils import TracerSpanContainer
 from tests.utils import call_program
 from tests.utils import request_token
 from tests.utils import snapshot_context as _snapshot_context
+
+
+code_to_pyc = getattr(
+    importlib._bootstrap_external, "_code_to_bytecode" if sys.version_info < (3, 7) else "_code_to_timestamp_pyc"
+)
 
 
 def pytest_configure(config):
@@ -115,31 +130,9 @@ def snapshot_context(request):
     return _snapshot
 
 
-# DEV: The dump_code_to_file function is adapted from the compile function in
-# the py_compile module of the Python standard library. It generates .pyc files
-# with the right format.
-if PY2:
-    import marshal
-    from py_compile import MAGIC
-    from py_compile import wr_long
-
-    def dump_code_to_file(code, file):
-        file.write(MAGIC)
-        wr_long(file, long(time.time()))  # noqa
-        marshal.dump(code, file)
-        file.flush()
-
-
-else:
-    import importlib
-
-    code_to_pyc = getattr(
-        importlib._bootstrap_external, "_code_to_bytecode" if sys.version_info < (3, 7) else "_code_to_timestamp_pyc"
-    )
-
-    def dump_code_to_file(code, file):
-        file.write(code_to_pyc(code, time.time(), len(code.co_code)))
-        file.flush()
+def dump_code_to_file(code, file):
+    file.write(code_to_pyc(code, time.time(), len(code.co_code)))
+    file.flush()
 
 
 def unwind_params(params):
@@ -301,9 +294,6 @@ def pytest_runtest_protocol(item):
         return True
 
 
-# source code fixtures
-
-
 def _run(cmd):
     return subprocess.check_output(cmd, shell=True)
 
@@ -393,3 +383,93 @@ def remote_config_worker():
     # we have 2 threads: main thread and telemetry thread. TODO: verify if that alive thread is a bug
     # TODO: this assert doesn't work in CI, threading.active_count() > 50
     # assert threading.active_count() == 2
+
+
+@pytest.fixture
+def telemetry_writer():
+    telemetry_writer = TelemetryWriter(is_periodic=False)
+    telemetry_writer.enable()
+    yield telemetry_writer
+
+
+@attr.s
+class TelemetryTestSession(object):
+    token = attr.ib(type=str)
+    telemetry_writer = attr.ib(type=TelemetryWriter)
+
+    def create_connection(self):
+        parsed = parse.urlparse(self.telemetry_writer._client._agent_url)
+        return httplib.HTTPConnection(parsed.hostname, parsed.port)
+
+    def _request(self, method, url):
+        # type: (str, str) -> Tuple[int, bytes]
+        conn = self.create_connection()
+        try:
+            conn.request(method, url)
+            r = conn.getresponse()
+            return r.status, r.read()
+        finally:
+            conn.close()
+
+    def clear(self):
+        status, _ = self._request("GET", "/test/session/clear?test_session_token=%s" % self.token)
+        if status != 200:
+            pytest.fail("Failed to clear session: %s" % self.token)
+        return True
+
+    def get_requests(self):
+        """Get a list of the requests sent to the test agent
+
+        Results are in reverse order by ``seq_id``
+        """
+        status, body = self._request("GET", "/test/session/requests?test_session_token=%s" % self.token)
+
+        if status != 200:
+            pytest.fail("Failed to fetch session requests: %s %s %s" % (self.create_connection(), status, self.token))
+        requests = json.loads(body.decode("utf-8"))
+        for req in requests:
+            body_str = base64.b64decode(req["body"]).decode("utf-8")
+            req["body"] = json.loads(body_str)
+
+        return sorted(requests, key=lambda r: r["body"]["seq_id"], reverse=True)
+
+    def get_events(self):
+        """Get a list of the event payloads sent to the test agent
+
+        Results are in reverse order by ``seq_id``
+        """
+        status, body = self._request("GET", "/test/session/apmtelemetry?test_session_token=%s" % self.token)
+        if status != 200:
+            pytest.fail("Failed to fetch session events: %s" % self.token)
+        return sorted(json.loads(body.decode("utf-8")), key=lambda e: e["seq_id"], reverse=True)
+
+
+@pytest.fixture
+def test_agent_session(telemetry_writer, request):
+    # type: (TelemetryWriter, Any) -> Generator[TelemetryTestSession, None, None]
+    token = request_token(request)
+    telemetry_writer._restart_sequence()
+    telemetry_writer._client._headers["X-Datadog-Test-Session-Token"] = token
+
+    # Also add a header to the environment for subprocesses test cases that might use snapshotting.
+    existing_headers = parse_tags_str(os.environ.get("_DD_TELEMETRY_WRITER_ADDITIONAL_HEADERS", ""))
+    existing_headers.update({"X-Datadog-Test-Session-Token": token})
+    os.environ["_DD_TELEMETRY_WRITER_ADDITIONAL_HEADERS"] = ",".join(
+        ["%s:%s" % (k, v) for k, v in existing_headers.items()]
+    )
+
+    requests = TelemetryTestSession(token=token, telemetry_writer=telemetry_writer)
+
+    conn = requests.create_connection()
+    try:
+        conn.request("GET", "/test/session/start?test_session_token=%s" % token)
+        conn.getresponse()
+    finally:
+        conn.close()
+
+    try:
+        yield requests
+    finally:
+        telemetry_writer.reset_queues()
+        del telemetry_writer._client._headers["X-Datadog-Test-Session-Token"]
+        del os.environ["_DD_TELEMETRY_WRITER_ADDITIONAL_HEADERS"]
