--- tests/runtime/conftest.py
@@ -1,3 +1,0 @@
-# Import helpful telemetry fixtures
-from ..telemetry.conftest import telemetry_writer  # noqa
-from ..telemetry.conftest import test_agent_session  # noqa
